refactor(sec-core): group skill ledger job
Group the Skill Ledger activation job under daemon/jobs so its registration,
notification handling, and processing helpers follow the existing job layout.

This is a structural move only; runtime behavior is unchanged.
