+++ src/agent-sec-core/agent-sec-cli/src/agent_sec_cli/daemon/jobs/registry.py
@@ -7,7 +7,7 @@ from agent_sec_cli.daemon.jobs.prompt_preload import (
     PromptModelPreloadJob,
     prompt_preload_enabled,
 )
-from agent_sec_cli.daemon.skill_ledger_activation import (
+from agent_sec_cli.daemon.jobs.skill_ledger import (
     SkillLedgerActivationJob,
 )
