+++ src/agent-sec-core/agent-sec-cli/src/agent_sec_cli/daemon/jobs/skill_ledger/__init__.py
@@ -1,0 +1,15 @@
+"""Stable daemon-facing exports for the Skill Ledger job."""
+
+from agent_sec_cli.daemon.jobs.skill_ledger.activation import (
+    METHOD_SKILLFS_NOTIFY_CHANGE,
+    SKILL_LEDGER_ACTIVATION_JOB,
+    SkillLedgerActivationJob,
+    skillfs_notify_method_spec,
+)
+
+__all__ = [
+    "METHOD_SKILLFS_NOTIFY_CHANGE",
+    "SKILL_LEDGER_ACTIVATION_JOB",
+    "SkillLedgerActivationJob",
+    "skillfs_notify_method_spec",
+]
